Refactor(Core): 공통 로직(BaseViewModel, LifecycleUtil)을 :core:common 모듈로 마이그레이션
- core/common 모듈 신설 및 Android Library 컨벤션 적용
- app 모듈의 BaseViewModel 및 LifecycleUtil을 core/common 모듈로 이동
- LifecycleUtil에서 사용되지 않는 Fragment 관련 코드 제거
- 패키지명 변경 및 관련 ViewModel의 임포트 경로 업데이트
- 멀티 모듈 리팩터링 계획서 업데이트
