+++ app/build.gradle.kts
@@ -27,6 +27,7 @@ dependencies {
     implementation(project(":domain"))
     implementation(project(":data"))
     implementation(project(":core:designsystem"))
+    implementation(project(":core:common"))
 
     implementation(platform(libs.androidx.compose.bom))
     androidTestImplementation(platform(libs.androidx.compose.bom))
