+++ app/src/main/java/com/gyleedev/githubsearch/ui/favorite/FavoriteViewModel.kt
@@ -2,7 +2,7 @@ package com.gyleedev.githubsearch.ui.favorite
 
 import androidx.lifecycle.viewModelScope
 import androidx.paging.cachedIn
-import com.gyleedev.githubsearch.core.BaseViewModel
+import com.gyleedev.githubsearch.core.common.BaseViewModel
 import com.gyleedev.githubsearch.domain.model.FilterStatus
 import com.gyleedev.githubsearch.domain.model.UserModel
 import com.gyleedev.githubsearch.domain.usecase.GetFavoritesUseCase
