+++ settings.gradle.kts
@@ -24,6 +24,7 @@ include(":feature:home")
 include(":feature:setting")
 include(":feature:favorite")
 include(":core")
+include(":core:common")
 include(":core:designsystem")
 include(":data")
 include(":domain")
